+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+FEATURES:
+* Add beta endpoints `ApplyToProjects`  and `RemoveFromProjects` to `VariableSets`.  Applying a variable set to a project will apply that variable set to all current and future workspaces in that project. 
+* Add beta endpoint `ListForProject` to `VariableSets` to list all variable sets applied to a project.
+
 ## Enhancements
 * Adds `ManageMembership` permission to team `OrganizationAccess` by @JarrettSpiker [#652](https://github.com/hashicorp/go-tfe/pull/652)
 * Adds `RotateKey` and `TrimKey` Admin endpoints by @mpminardi [#666](https://github.com/hashicorp/go-tfe/pull/666)